--- .env.example
@@ -1,11 +1,0 @@
-AWS_KEY_1=키를_적어주세요
-AWS_SECRET_1=스크릿_키를_적어주세요
-
-AWS_KEY_2=키를_적어주세요
-AWS_SECRET_2=스크릿_키를_적어주세요
-
-AWS_KEY_3=키를_적어주세요
-AWS_SECRET_3=스크릿_키를_적어주세요
-
-AWS_KEY_4=키를_적어주세요
-AWS_SECRET_4=스크릿_키를_적어주세요
