+++ app/step/4/page.js
@@ -38,10 +38,7 @@ export default async function Step4() {
                 <ClipPreview videoClientProps={videoClientProps} />
             </div>
             <LetterButton urlParams={{ ...encode, to: videoClientProps.to }} />
-            <Download
-                encode={encode}
-                transformedVideoProps={transformedVideoProps}
-            />
+            <Download encode={encode} />
         </div>
     )
 }
