+++ constant/path.js
@@ -1,7 +1,7 @@
 const pathBase =
     process.env.NODE_ENV === "development"
         ? "http://localhost:3000"
-        : "https://thanksclip.vercel.app"
+        : "https://thxtest.vercel.app"
 
 //!TODO: 배포 주소로 변경하기
