+++ package.json
@@ -26,6 +26,7 @@
         "@remotion/cli": "3.3.27",
         "@remotion/lambda": "3.3.27",
         "@remotion/player": "3.3.27",
+        "@vercel/og": "^0.0.27",
         "cookies-next": "^2.1.1",
         "jotai": "^1.13.1",
         "next": "^13.1.2",
@@ -57,7 +58,6 @@
         "postcss-loader": "7.0.2",
         "postcss-preset-env": "7.8.3",
         "prettier": "^2.8.3",
-        "satori": "^0.1.1",
         "sharp": "^0.31.3",
         "style-loader": "^3.3.1",
         "tailwindcss": "^3.2.4",
